+++ src/broadcasts.rs
@@ -288,7 +288,7 @@ pub mod types {
         }
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct UpdateBroadcastResponse {
         /// Unique identifier for the updated broadcast.
         pub id: BroadcastId,
@@ -296,7 +296,7 @@ pub mod types {
 
     crate::define_id_type!(BroadcastId);
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct CreateBroadcastResponse {
         /// The ID of the created broadcast.
         pub id: BroadcastId,
@@ -330,14 +330,14 @@ pub mod types {
         }
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct SendBroadcastResponse {
         /// The ID of the sent broadcast.
         pub id: BroadcastId,
     }
 
     #[must_use]
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct Broadcast {
         pub id: BroadcastId,
         pub name: String,
@@ -354,7 +354,7 @@ pub mod types {
         pub html: Option<String>,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct RemoveBroadcastResponse {
         /// The ID of the broadcast.
         #[allow(dead_code)]
@@ -380,6 +380,7 @@ mod test {
 
     #[tokio_shared_rt::test(shared = true)]
     #[cfg(not(feature = "blocking"))]
+    #[ignore = "Can no longer send broadcasts from the resend.dev domain"]
     async fn create_send_broadcast() -> DebugResult<()> {
         let resend = &*CLIENT;
         std::thread::sleep(std::time::Duration::from_secs(1));
@@ -424,6 +425,7 @@ mod test {
 
     #[tokio_shared_rt::test(shared = true)]
     #[cfg(not(feature = "blocking"))]
+    #[ignore = "Can no longer send broadcasts from the resend.dev domain"]
     async fn list_get_broadcast() -> DebugResult<()> {
         let resend = &*CLIENT;
         std::thread::sleep(std::time::Duration::from_secs(1));
@@ -459,7 +461,7 @@ mod test {
 
     #[tokio_shared_rt::test(shared = true)]
     #[cfg(not(feature = "blocking"))]
-    #[track_caller]
+    #[ignore = "Can no longer send broadcasts from the resend.dev domain"]
     async fn update_broadcast() -> DebugResult<()> {
         let resend = &*CLIENT;
         std::thread::sleep(std::time::Duration::from_secs(1));