+++ src/domains.rs
@@ -207,7 +207,7 @@ pub mod types {
         ApNorthEast1,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct DomainSpfRecord {
         /// The name of the record.
         pub name: String,
@@ -215,7 +215,7 @@ pub mod types {
         pub value: String,
         /// The type of record.
         #[serde(rename = "type")]
-        pub d_type: SpfRecordType,
+        pub r#type: SpfRecordType,
         /// The time to live for the record.
         pub ttl: String,
         /// The status of the record.
@@ -226,7 +226,7 @@ pub mod types {
         pub proxy_status: Option<ProxyStatus>,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct DomainDkimRecord {
         /// The name of the record.
         pub name: String,
@@ -234,7 +234,7 @@ pub mod types {
         pub value: String,
         /// The type of record.
         #[serde(rename = "type")]
-        pub d_type: DkimRecordType,
+        pub r#type: DkimRecordType,
         /// The time to live for the record.
         pub ttl: String,
         /// The status of the record.
@@ -245,7 +245,7 @@ pub mod types {
         pub proxy_status: Option<ProxyStatus>,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct ReceivingRecord {
         /// The name of the record.
         pub name: String,
@@ -253,7 +253,7 @@ pub mod types {
         pub value: String,
         /// The type of record.
         #[serde(rename = "type")]
-        pub d_type: ReceivingRecordType,
+        pub r#type: ReceivingRecordType,
         /// The time to live for the record.
         pub ttl: String,
         /// The status of the record.
@@ -262,37 +262,36 @@ pub mod types {
         pub priority: i32,
     }
 
-    #[derive(Debug, Copy, Clone, Deserialize)]
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
     pub enum ReceivingRecordType {
         #[allow(clippy::upper_case_acronyms)]
         MX,
     }
 
-    #[derive(Debug, Copy, Clone, Deserialize)]
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
     pub enum ProxyStatus {
         Enable,
         Disable,
     }
 
-    #[derive(Debug, Copy, Clone, Deserialize)]
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
     pub enum DomainStatus {
         Pending,
         Verified,
         Failed,
-        #[serde(rename = "temporary_failure")]
         TemporaryFailure,
-        #[serde(rename = "not_started")]
         NotStarted,
     }
 
-    #[derive(Debug, Copy, Clone, Deserialize)]
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
     pub enum SpfRecordType {
         MX,
         #[allow(clippy::upper_case_acronyms)]
         TXT,
     }
 
-    #[derive(Debug, Copy, Clone, Deserialize)]
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
     pub enum DkimRecordType {
         #[allow(clippy::upper_case_acronyms)]
         CNAME,
@@ -301,20 +300,20 @@ pub mod types {
     }
 
     /// Individual [`Domain`] record.
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     #[serde(tag = "record")]
     pub enum DomainRecord {
         #[serde(rename = "SPF")]
         DomainSpfRecord(DomainSpfRecord),
         #[serde(rename = "DKIM")]
         DomainDkimRecord(DomainDkimRecord),
-        #[serde(rename = "Receiving")]
+        #[serde(rename = "Receiving MX")]
         ReceivingRecord(ReceivingRecord),
     }
 
     /// Details of an existing domain.
     #[must_use]
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct Domain {
         /// The ID of the domain.
         pub id: DomainId,
@@ -332,7 +331,7 @@ pub mod types {
         pub records: Option<Vec<DomainRecord>>,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct VerifyDomainResponse {
         /// The ID of the domain.
         #[allow(dead_code)]
@@ -382,13 +381,13 @@ pub mod types {
         }
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct UpdateDomainResponse {
         /// The ID of the updated domain.
         pub id: DomainId,
     }
 
-    #[derive(Debug, Clone, Deserialize)]
+    #[derive(Debug, Clone, Serialize, Deserialize)]
     pub struct DeleteDomainResponse {
         /// The ID of the domain.
         pub id: DomainId,