+++ src/error.rs
@@ -1,11 +1,11 @@
 #[allow(unreachable_pub)]
 pub mod types {
-    use serde::Deserialize;
+    use serde::{Deserialize, Serialize};
 
     /// Error returned as a response.
     ///
     /// <https://resend.com/docs/api-reference/errors>
-    #[derive(Debug, Clone, Deserialize, thiserror::Error)]
+    #[derive(Debug, Clone, Serialize, Deserialize, thiserror::Error)]
     #[error("{name}: {message}")]
     pub struct ErrorResponse {
         #[serde(rename = "statusCode")]
@@ -85,28 +85,11 @@ pub mod types {
         ///
         /// - `validation_error`
         ///
-        /// You can only send testing emails to your own email address (`youremail@domain.com`).
-        ///
-        /// To send emails to other recipients, please verify a domain at resend.com/domains, and
-        /// change the from address to an email using this domain.
-        ///
-        /// In [Resend's Domain page], add and verify a domain for
-        /// which you have DNS access. This allows you to send emails to addresses beyond your own.
-        ///
-        /// [Learn more about resolving this error](https://resend.com/docs/knowledge-base/403-error-resend-dev-domain).
-        ValidationError403Email,
-
-        /// 403 Forbidden.
-        ///
-        /// - `validation_error`
-        ///
-        /// The `domain.com` domain is not verified. Please, add and verify your domain.
-        ///
-        /// Make sure the domain in your API request's `from` field matches a domain you've
-        /// verified in Resend. Update your API request to use your verified domain, or add and
-        /// verify the domain you're trying to use.
-        /// [Learn more about resolving this error](https://resend.com/docs/knowledge-base/403-error-domain-mismatch).
-        ValidationError403Domain,
+        /// One of the following:
+        /// - <https://resend.com/docs/api-reference/errors#validation_error-2>
+        /// - <https://resend.com/docs/api-reference/errors#validation_error-3>
+        /// - <https://resend.com/docs/api-reference/errors#validation_error-4>
+        ValidationError403,
 
         /// 404 Not Found.
         ///
@@ -277,13 +260,7 @@ pub mod types {
                     400 => Self::ValidationError400,
                     // This is a bit silly, since we have 2 validation errors with the same error
                     // code, we need to differentiate between them based on the message.
-                    403 => {
-                        if value.message.starts_with("You") {
-                            Self::ValidationError403Email
-                        } else {
-                            Self::ValidationError403Domain
-                        }
-                    }
+                    403 => Self::ValidationError403,
                     _ => Self::Unrecognized,
                 };
             }
@@ -384,7 +361,7 @@ mod test {
         }
 
         // Expected is actually one less than what we have because of the `Unrecognized` variant.
-        let expected = expected.len() + 1;
+        let expected = expected.len() - 1;
 
         assert_eq!(actual, expected);
     }