+++ src/events.rs
@@ -5,9 +5,14 @@
 
 #![allow(dead_code)]
 
+use std::collections::HashMap;
+
 use serde::{Deserialize, Serialize};
 
-use crate::{Result, types::Domain};
+use crate::{
+    Result,
+    types::{BroadcastId, Domain, EmailId, InboundAttachment, SegmentId, TemplateId},
+};
 
 /// Parses a JSON event into an [`Event`].
 /// ## Example
@@ -18,15 +23,20 @@ use crate::{Result, types::Domain};
 /// let data = r#"
 ///   {
 ///     "type": "email.sent",
-///     "created_at": "2024-11-23T15:53:07.839Z",
+///     "created_at": "2024-02-22T23:41:12.126Z",
 ///     "data": {
-///         "created_at": "2024-11-23 15:53:07.743225+00",
-///         "email_id": "9a148e6d-d79f-43cb-8022-22320546e1db",
-///         "from": "Acme <onboarding@resend.dev>",
-///         "subject": "hello world",
-///         "to": ["delivered@resend.dev"]
+///       "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
+///       "created_at": "2024-02-22T23:41:11.894719+00:00",
+///       "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+///       "from": "Acme <onboarding@resend.dev>",
+///       "to": ["delivered@resend.dev"],
+///       "subject": "Sending this example",
+///       "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+///       "tags": {
+///         "category": "confirm_email"
+///       }
 ///     }
-///   }"#;
+/// }"#;
 ///
 ///  let parsed: Result<Event, resend_rs::Error> = try_parse_event(data);
 /// ```
@@ -49,8 +59,9 @@ pub fn try_parse_event_type(data: &str) -> Result<EventType> {
 }
 
 /// Represents any [Resend Event](https://resend.com/docs/dashboard/webhooks/event-types).
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
+#[allow(clippy::large_enum_variant)]
 pub enum Event {
     EmailEvent(EmailEvent),
     ContactEvent(ContactEvent),
@@ -84,7 +95,7 @@ impl From<DomainEventType> for EventType {
     }
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EmailEvent {
     #[serde(rename = "type")]
     pub r#type: EmailEventType,
@@ -92,7 +103,7 @@ pub struct EmailEvent {
     pub data: EmailBody,
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ContactEvent {
     #[serde(rename = "type")]
     pub r#type: ContactEventType,
@@ -100,7 +111,7 @@ pub struct ContactEvent {
     pub data: ContactBody,
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DomainEvent {
     #[serde(rename = "type")]
     pub r#type: DomainEventType,
@@ -157,18 +168,68 @@ pub enum DomainEventType {
     DomainDeleted,
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EmailBody {
+    pub broadcast_id: Option<BroadcastId>,
     pub created_at: String,
-    pub email_id: String,
+    pub email_id: EmailId,
     pub from: String,
     pub to: Vec<String>,
-    pub click: Option<Click>,
     pub subject: String,
+    pub template_id: Option<TemplateId>,
+
+    #[serde(flatten)]
+    pub received: Option<Received>,
+    pub click: Option<Click>,
+    pub bounce: Option<Bounce>,
+    pub failed: Option<Failed>,
+    pub suppressed: Option<Suppressed>,
+
+    #[serde(default)]
+    pub tags: HashMap<String, String>,
+}
+
+/// Extra data only populated in [`EmailEventType::EmailSuppressed`] events.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Suppressed {
+    pub message: String,
+    #[serde(rename = "type")]
+    pub r#type: String,
+}
+
+/// Extra data only populated in [`EmailEventType::EmailReceived`] events.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Received {
+    pub bcc: Vec<String>,
+    pub cc: Vec<String>,
+    pub message_id: String,
+    pub attachments: Vec<InboundAttachment>,
+}
+
+/// Extra data only populated in [`EmailEventType::EmailFailed`] events.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Failed {
+    pub reason: String,
+}
+
+/// Extra data only populated in [`EmailEventType::EmailBounced`] events.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Bounce {
+    pub message: String,
+    #[serde(rename = "subType")]
+    pub sub_type: BounceType,
+    #[serde(rename = "type")]
+    pub r#type: String,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+pub enum BounceType {
+    Suppressed,
+    MessageRejected,
 }
 
 /// Extra data only populated in [`EmailEventType::EmailClicked`] events.
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Click {
     #[serde(rename = "ipAddress")]
     pub ip_address: String,
@@ -178,15 +239,16 @@ pub struct Click {
     pub user_agent: String,
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ContactBody {
     pub id: String,
     pub audience_id: String,
+    pub segment_ids: Vec<SegmentId>,
     pub created_at: String,
     pub updated_at: String,
     pub email: String,
-    pub first_name: String,
-    pub last_name: String,
+    pub first_name: Option<String>,
+    pub last_name: Option<String>,
     pub unsubscribed: bool,
 }
 
@@ -206,13 +268,18 @@ mod test {
         let data = r#"
     {
       "type": "email.sent",
-      "created_at": "2024-11-23T15:53:07.839Z",
+      "created_at": "2024-02-22T23:41:12.126Z",
       "data": {
-          "created_at": "2024-11-23 15:53:07.743225+00",
-          "email_id": "9a148e6d-d79f-43cb-8022-22320546e1db",
-          "from": "Acme <onboarding@resend.dev>",
-          "subject": "hello world",
-          "to": ["delivered@resend.dev"]
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
+        "created_at": "2024-02-22T23:41:11.894719+00:00",
+        "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+        "from": "Acme <onboarding@resend.dev>",
+        "to": ["delivered@resend.dev"],
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -234,11 +301,16 @@ mod test {
       "type": "email.delivered",
       "created_at": "2024-02-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-02-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
         "to": ["delivered@resend.dev"],
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -260,11 +332,16 @@ mod test {
       "type": "email.delivery_delayed",
       "created_at": "2024-02-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-02-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
         "to": ["delivered@resend.dev"],
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -289,11 +366,16 @@ mod test {
       "type": "email.complained",
       "created_at": "2024-02-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-02-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
         "to": ["delivered@resend.dev"],
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -318,11 +400,21 @@ mod test {
       "type": "email.bounced",
       "created_at": "2024-11-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-11-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
         "to": ["delivered@resend.dev"],
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "bounce": {
+          "message": "The recipient's email address is on the suppression list because it has a recent history of producing hard bounces.",
+          "subType": "Suppressed",
+          "type": "Permanent"
+        },
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -332,6 +424,7 @@ mod test {
 
         if let Event::EmailEvent(email_event) = parsed {
             assert!(matches!(email_event.r#type, EmailEventType::EmailBounced));
+            assert!(email_event.data.bounce.is_some());
         } else {
             panic!("Wrong parsing");
         }
@@ -344,11 +437,16 @@ mod test {
       "type": "email.opened",
       "created_at": "2024-02-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-02-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
         "to": ["delivered@resend.dev"],
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -370,6 +468,7 @@ mod test {
       "type": "email.clicked",
       "created_at": "2024-11-22T23:41:12.126Z",
       "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
         "created_at": "2024-11-22T23:41:11.894719+00:00",
         "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
         "from": "Acme <onboarding@resend.dev>",
@@ -380,7 +479,11 @@ mod test {
           "timestamp": "2024-11-24T05:00:57.163Z",
           "userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.1 Safari/605.1.15"
         },
-        "subject": "Sending this example"
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
       }
     }"#;
 
@@ -391,6 +494,162 @@ mod test {
         if let Event::EmailEvent(email_event) = parsed {
             assert!(matches!(email_event.r#type, EmailEventType::EmailClicked));
             assert!(email_event.data.click.is_some());
+            assert!(!email_event.data.tags.is_empty());
+        } else {
+            panic!("Wrong parsing");
+        }
+    }
+
+    #[test]
+    fn email_failed() {
+        let data = r#"
+    {
+      "type": "email.failed",
+      "created_at": "2024-11-22T23:41:12.126Z",
+      "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
+        "created_at": "2024-11-22T23:41:11.894719+00:00",
+        "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+        "from": "Acme <onboarding@resend.dev>",
+        "to": ["delivered@resend.dev"],
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "failed": {
+          "reason": "reached_daily_quota"
+        },
+        "tags": {
+          "category": "confirm_email"
+        }
+      }
+    }"#;
+
+        let parsed = try_parse_event(data);
+        assert!(parsed.is_ok());
+        let parsed = parsed.unwrap();
+
+        if let Event::EmailEvent(email_event) = parsed {
+            assert!(matches!(email_event.r#type, EmailEventType::EmailFailed));
+            assert!(email_event.data.failed.is_some());
+            assert!(!email_event.data.tags.is_empty());
+        } else {
+            panic!("Wrong parsing");
+        }
+    }
+
+    #[test]
+    fn email_received() {
+        let data = r#"
+    {
+      "type": "email.received",
+      "created_at": "2024-02-22T23:41:12.126Z",
+      "data": {
+        "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+        "created_at": "2024-02-22T23:41:11.894719+00:00",
+        "from": "Acme <onboarding@resend.dev>",
+        "to": ["delivered@resend.dev"],
+        "bcc": [],
+        "cc": [],
+        "message_id": "<example+123>",
+        "subject": "Sending this example",
+        "attachments": [
+          {
+            "id": "2a0c9ce0-3112-4728-976e-47ddcd16a318",
+            "filename": "avatar.png",
+            "content_type": "image/png",
+            "content_disposition": "inline",
+            "content_id": "img001"
+          }
+        ]
+      }
+    }"#;
+
+        let parsed = try_parse_event(data);
+        assert!(parsed.is_ok());
+        let parsed = parsed.unwrap();
+
+        if let Event::EmailEvent(email_event) = parsed {
+            assert!(matches!(email_event.r#type, EmailEventType::EmailReceived));
+            assert!(email_event.data.received.is_some());
+            assert!(email_event.data.tags.is_empty());
+
+            let received = email_event.data.received.unwrap();
+            assert!(received.attachments.len() == 1);
+            assert!(received.cc.is_empty());
+            assert!(received.bcc.is_empty());
+        } else {
+            panic!("Wrong parsing");
+        }
+    }
+
+    #[test]
+    fn email_scheduled() {
+        let data = r#"
+    {
+      "type": "email.scheduled",
+      "created_at": "2024-02-22T23:41:12.126Z",
+      "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
+        "created_at": "2024-02-22T23:41:11.894719+00:00",
+        "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+        "from": "Acme <onboarding@resend.dev>",
+        "to": ["delivered@resend.dev"],
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "tags": {
+          "category": "confirm_email"
+        }
+      }
+    }"#;
+
+        let parsed = try_parse_event(data);
+        assert!(parsed.is_ok());
+        let parsed = parsed.unwrap();
+
+        if let Event::EmailEvent(email_event) = parsed {
+            assert!(matches!(email_event.r#type, EmailEventType::EmailScheduled));
+            assert!(email_event.data.received.is_none());
+            assert!(!email_event.data.tags.is_empty());
+        } else {
+            panic!("Wrong parsing");
+        }
+    }
+
+    #[test]
+    fn email_suppressed() {
+        let data = r#"
+    {
+      "type": "email.suppressed",
+      "created_at": "2024-11-22T23:41:12.126Z",
+      "data": {
+        "broadcast_id": "8b146471-e88e-4322-86af-016cd36fd216",
+        "created_at": "2024-11-22T23:41:11.894719+00:00",
+        "email_id": "56761188-7520-42d8-8898-ff6fc54ce618",
+        "from": "Acme <onboarding@resend.dev>",
+        "to": ["delivered@resend.dev"],
+        "subject": "Sending this example",
+        "template_id": "43f68331-0622-4e15-8202-246a0388854b",
+        "suppressed": {
+          "message": "Resend has suppressed sending to this address because it is on the account-level suppression list. This does not count toward your bounce rate metric",
+          "type": "OnAccountSuppressionList"
+        },
+        "tags": {
+          "category": "confirm_email"
+        }
+      }
+    }"#;
+
+        let parsed = try_parse_event(data);
+        assert!(parsed.is_ok());
+        let parsed = parsed.unwrap();
+
+        if let Event::EmailEvent(email_event) = parsed {
+            assert!(matches!(
+                email_event.r#type,
+                EmailEventType::EmailSuppressed
+            ));
+            assert!(email_event.data.received.is_none());
+            assert!(!email_event.data.tags.is_empty());
+            assert!(email_event.data.suppressed.is_some());
         } else {
             panic!("Wrong parsing");
         }
@@ -405,6 +664,7 @@ mod test {
       "data": {
         "id": "e169aa45-1ecf-4183-9955-b1499d5701d3",
         "audience_id": "78261eea-8f8b-4381-83c6-79fa7120f1cf",
+        "segment_ids": ["78261eea-8f8b-4381-83c6-79fa7120f1cf"],
         "created_at": "2024-11-17T19:32:22.980Z",
         "updated_at": "2024-11-17T19:32:22.980Z",
         "email": "steve.wozniak@gmail.com",
@@ -423,6 +683,7 @@ mod test {
                 contact_event.r#type,
                 ContactEventType::ContactCreated
             ));
+            assert!(contact_event.data.segment_ids.len() == 1);
         } else {
             panic!("Wrong parsing");
         }
@@ -437,6 +698,7 @@ mod test {
       "data": {
         "id": "e169aa45-1ecf-4183-9955-b1499d5701d3",
         "audience_id": "78261eea-8f8b-4381-83c6-79fa7120f1cf",
+        "segment_ids": ["78261eea-8f8b-4381-83c6-79fa7120f1cf"],
         "created_at": "2024-10-10T15:11:94.110Z",
         "updated_at": "2024-10-11T23:47:56.678Z",
         "email": "steve.wozniak@gmail.com",
@@ -455,6 +717,7 @@ mod test {
                 contact_event.r#type,
                 ContactEventType::ContactUpdated
             ));
+            assert!(contact_event.data.segment_ids.len() == 1);
         } else {
             panic!("Wrong parsing");
         }
@@ -469,6 +732,7 @@ mod test {
       "data": {
         "id": "e169aa45-1ecf-4183-9955-b1499d5701d3",
         "audience_id": "78261eea-8f8b-4381-83c6-79fa7120f1cf",
+        "segment_ids": ["78261eea-8f8b-4381-83c6-79fa7120f1cf"],
         "created_at": "2024-11-10T15:11:94.110Z",
         "updated_at": "2024-11-17T19:32:22.980Z",
         "email": "steve.wozniak@gmail.com",
@@ -487,6 +751,7 @@ mod test {
                 contact_event.r#type,
                 ContactEventType::ContactDeleted
             ));
+            assert!(contact_event.data.segment_ids.len() == 1);
         } else {
             panic!("Wrong parsing");
         }
@@ -502,7 +767,6 @@ mod test {
         "id": "d91cd9bd-1176-453e-8fc1-35364d380206",
         "name": "example.com",
         "status": "not_started",
-        "capability": "send",
         "created_at": "2024-04-26T20:21:26.347412+00:00",
         "region": "us-east-1",
         "records": [
@@ -560,7 +824,6 @@ mod test {
         "id": "d91cd9bd-1176-453e-8fc1-35364d380206",
         "name": "example.com",
         "status": "not_started",
-        "capability": "send",
         "created_at": "2024-04-26T20:21:26.347412+00:00",
         "region": "us-east-1",
         "records": [
@@ -588,6 +851,15 @@ mod test {
             "type": "TXT",
             "status": "not_started",
             "ttl": "Auto"
+          },
+          {
+            "name": "inbound.yourdomain.tld",
+            "priority": 10,
+            "record": "Receiving MX",
+            "status": "pending",
+            "ttl": "Auto",
+            "type": "MX",
+            "value": "inbound-smtp.us-east-1.amazonaws.com"
           }
         ]
       }
@@ -602,7 +874,7 @@ mod test {
                 domain_event.r#type,
                 DomainEventType::DomainUpdated
             ));
-            assert!(domain_event.data.records.is_some_and(|r| r.len() == 3));
+            assert!(domain_event.data.records.is_some_and(|r| r.len() == 4));
         } else {
             panic!("Wrong parsing");
         }
@@ -618,7 +890,6 @@ mod test {
         "id": "d91cd9bd-1176-453e-8fc1-35364d380206",
         "name": "example.com",
         "status": "not_started",
-        "capability": "send",
         "created_at": "2024-04-26T20:21:26.347412+00:00",
         "region": "us-east-1",
         "records": [